requirements.txt
asgiref==3.4.1
Django==3.2.5
django-phonenumber-field==5.2.0
djongo==1.3.6
phonenumbers==8.12.28
pymongo==3.12.0
pytz==2021.1
sqlparse==0.2.4
